pyproject.toml
[project]
name = "ksf-restful"
version = "0.1.0"
description = "一个基于Flask-RESTful的完整REST API项目框架"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = "3.12"
dependencies = [
    "Flask>=2.3.3,<3.0.0",
    "flask-restx>=1.3.0,<2.0.0",
    "Flask-SQLAlchemy>=3.0.5,<4.0.0",
    "Flask-Migrate>=4.0.5,<5.0.0",
    "Flask-CORS>=4.0.0,<5.0.0",
    "Flask-Limiter>=3.5.0,<4.0.0",
    "marshmallow>=3.20.1,<4.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "gunicorn>=21.2.0,<22.0.0",
    "click>=8.0.0,<9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.2,<8.0.0",
    "pytest-flask>=1.3.0,<2.0.0",
    "coverage>=7.3.2,<8.0.0",
    "black>=23.9.1,<24.0.0",
    "flake8>=6.1.0,<7.0.0",
    "mypy>=1.5.1,<2.0.0",
    "pre-commit>=3.4.0,<4.0.0",
    "isort>=5.12.0,<6.0.0",
]
test = [
    "pytest>=7.4.2,<8.0.0",
    "pytest-flask>=1.3.0,<2.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "factory-boy>=3.3.0,<4.0.0",
]

[project.scripts]
ksf-restful = "run:app"

[tool.poetry]
packages = [
    {include = "app"}
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    "venv",
    ".mypy_cache",
    ".pytest_cache"
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true 